refactor: 💡 remove deprecated file and functions
